Refactor code, remove unused codes
