+++ ide/h1st/src/browser/visualization/components/graph/consts.ts
@@ -1,0 +1,24 @@
+export const NODE_KEY = "id";
+export const EMPTY_TYPE = "empty";
+export const CUSTOM_EMPTY_TYPE = "customEmpty"; // Empty node type
+export const START_TYPE = "start";
+export const END_TYPE = "end";
+export const POLY_TYPE = "poly";
+export const SPECIAL_TYPE = "special";
+export const ACTION_TYPE = "action";
+export const SPECIAL_CHILD_SUBTYPE = "specialChild";
+export const EMPTY_EDGE_TYPE = "emptyEdge";
+export const SPECIAL_EDGE_TYPE = "specialEdge";
+
+export const nodeTypes = [
+  EMPTY_TYPE,
+  CUSTOM_EMPTY_TYPE,
+  POLY_TYPE,
+  SPECIAL_TYPE,
+  ACTION_TYPE,
+  START_TYPE,
+  END_TYPE,
+];
+export const edgeTypes = [EMPTY_EDGE_TYPE, SPECIAL_EDGE_TYPE];
+
+export const nodeSubTypes = [SPECIAL_CHILD_SUBTYPE];
