+++ ide/h1st/src/browser/visualization/containers/GraphContainer.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { Provider } from "react-redux";
-import Graph from "../components/graph";
+import Graph from "../components/graph/Graph";
 
 export const GraphContainer = (props: { store: any }) => {
   return (
